move options to arg.py
